requirements.txt
matplotlib>=3.3.4
numpy>=1.19.5
opencv-python>=4.5.5
pandas>=1.1.5
Pillow>=8.4.0
scipy>=1.5.4
torch>=1.10.1
torchvision>=0.11.2
git+https://github.com/elliottzheng/face-detection